compute `<style>` and `</style>` positions once
This allows us to compute all the positions once and cache it per file
contents. This is useful because otherwise we have to check the exact
same file over and over and over again, but just for different
positions.

This also just computes the ranges of start and end positions. This will
allow us to simply check if the candidate is between any of 2 pair
points to know if we are in a `<style>…</style>` or not.